feat(web)!: target TypeScript, keep a JavaScript workaround aside
Colibri targets TypeScript: @Synced() is a TypeScript decorator, and the
documentation, samples and tests all assume a TypeScript project.

- remove the plain-JS sample ports and their samples/*-js npm scripts
- drop the "SyncModel in plain JavaScript" section from the web README and
  state the TypeScript target under Installation instead
- record the change as breaking in the changelog

For projects that are tied to plain JavaScript, the decorator-free SyncModel
workaround lives in colibri-web/docs/js-workaround/. It is unlinked from the
main documentation, excluded from the published package, and notes that a
workaround is not a feature: untested, dependent on members that are public
only because @Synced() needs them, and liable to break in future releases.
