+++ colibri-web/package.json
@@ -29,13 +29,9 @@
         "prepublishOnly": "npm run build",
         "publish": "npm run build && npm publish --access public",
         "samples/broadcast": "tsx samples/broadcast.ts",
-        "samples/broadcast-js": "npm run build && node samples/broadcast.js",
         "samples/model-sync": "tsx samples/model-sync.ts",
-        "samples/model-sync-js": "npm run build && node samples/model-sync.js",
         "samples/remote-logging": "tsx samples/remote-logging.ts",
-        "samples/remote-logging-js": "npm run build && node samples/remote-logging.js",
         "samples/rest-api": "tsx samples/rest-api.ts",
-        "samples/rest-api-js": "npm run build && node samples/rest-api.js",
         "test": "vitest run",
         "test:watch": "vitest",
         "test:coverage": "vitest run --coverage",
